pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "waywo"
version = "0.1.0"
description = "An searchable index of 'What are you working on?' projects from Hacker News"
readme = "README.md"
license = {text = "Apache 2.0"}
authors = [
    {name = "Brian", email = "brian@waywo.com"}
]
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "redis>=5.0.0",
    "httpx>=0.24.0",
    "celery[redis]>=5.5.3",
    "flower>=2.0.1",
    "python-dateutil>=2.8.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "numpy>=1.24.0",
    "openai==1.109.0",
    "python-multipart>=0.0.20",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "langgraph>=0.2.0",
    "watchdog>=4.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "llama-index-core>=0.12.0",
    "llama-index-llms-openai-like>=0.3.0",
    "sqliteai-vector>=0.1.0",
    "nest-asyncio>=1.6.0",
    # OpenTelemetry + Phoenix observability
    "openinference-instrumentation-llama-index>=3.0.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    # Workflow visualization
    "llama-index-utils-workflow>=0.9.0",
    # Screenshot capture
    "playwright>=1.40.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "eralchemy>=1.6.0",
    "graphviz>=0.20",
]
notebook = [
    "jupyterlab>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/briancaffey/waywo"
Repository = "https://github.com/briancaffey/waywo.git"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
testpaths = ["src/test"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings"
]
markers = [
    "route: API route endpoint tests",
    "db: database operation tests",
    "client: external service client tests",
    "worker: Celery worker task tests",
    "integration: end-to-end workflow tests"
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/test/*",
    "*/__pycache__/*",
    "*/migrations/*",
    "*/venv/*",
    "*/.venv/*",
    "*/env/*",
    "*/site-packages/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
precision = 2

